+++ pos_debt_notebook/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from . import models
 from . import report
+from . import wizard
 
 from openerp import SUPERUSER_ID
 from openerp import api