+++ pos_debt_notebook/models.py
@@ -1,7 +1,11 @@
 # -*- coding: utf-8 -*-
 from odoo import models, fields, api, _
 from odoo.exceptions import UserError
+from datetime import datetime
+from pytz import timezone
+import pytz
 import odoo.addons.decimal_precision as dp
+from odoo.tools import DEFAULT_SERVER_DATETIME_FORMAT
 
 
 class ResPartner(models.Model):
@@ -23,6 +27,30 @@ class ResPartner(models.Model):
         for r in self:
             r.debt = -res_index[r.id]['balance']
             r.credit_balance = -r.debt
+
+    @api.multi
+    def _compute_debt_company(self):
+        partners = self.filtered(lambda r: len(r.child_ids))
+        if not partners:
+            return
+        domain = [('partner_id', 'in', partners.ids + partners.mapped('child_ids').ids)]
+        fields = ['partner_id', 'balance']
+        res = self.env['report.pos.debt'].read_group(
+            domain,
+            fields,
+            'partner_id')
+
+        res_index = dict((id, 0) for id in partners.ids)
+        for data in res:
+            id = data['partner_id'][0]
+            balance = data['balance']
+            for r in partners:
+                if id == r.id or id in r.child_ids.ids:
+                    res_index[r.id] += balance
+
+        for r in partners:
+            r.debt_company = -res_index[r.id]
+            r.credit_balance_company = -r.debt_company
 
     @api.model
     def _default_debt_limit(self):
@@ -63,6 +91,8 @@ class ResPartner(models.Model):
                     fields=fields,
                     limit=limit,
                 )
+                for r2 in records:
+                    r2['date'] = self._get_date_formats(r2['date'])
                 data['history'] = records
             data['records_count'] = self.env['report.pos.debt'].search_count(domain)
             data['partner_id'] = r.id
@@ -72,17 +102,38 @@ class ResPartner(models.Model):
 
     debt = fields.Float(
         compute='_compute_debt', string='Debt', readonly=True,
-        digits=dp.get_precision('Account'), help='This debt value for only current company')
+        digits=dp.get_precision('Account'), help='Debt of this partner only.')
     credit_balance = fields.Float(
         compute='_compute_debt', string='Credit', readonly=True,
-        digits=dp.get_precision('Account'), help='This credit balance value for only current company')
+        digits=dp.get_precision('Account'), help='Credit balance of this partner only.')
+    debt_company = fields.Float(
+        compute='_compute_debt_company', string='Total Debt', readonly=True,
+        digits=dp.get_precision('Account'), help='Debt value of this company (including its contacts)')
+    credit_balance_company = fields.Float(
+        compute='_compute_debt_company', string='Total Credit', readonly=True,
+        digits=dp.get_precision('Account'), help='Credit balance of this company (including its contacts)')
     debt_type = fields.Selection(compute='_compute_debt_type', selection=[
         ('debt', 'Display Debt'),
         ('credit', 'Display Credit')
     ])
     debt_limit = fields.Float(
         string='Max Debt', digits=dp.get_precision('Account'), default=_default_debt_limit,
-        help='The customer is not allowed to have a debt more than this value')
+        help='The partner is not allowed to have a debt more than this value')
+
+    def _get_date_formats(self, report):
+
+        lang_code = self.env.user.lang or 'en_US'
+        lang = self.env['res.lang']._lang_get(lang_code)
+        date_format = lang.date_format
+        time_format = lang.time_format
+        fmt = date_format + " " + time_format
+
+        server_date = datetime.strptime(report, DEFAULT_SERVER_DATETIME_FORMAT)
+        utc_tz = pytz.utc.localize(server_date, is_dst=False)
+        user_tz = timezone(self.env.user.tz)
+        final = utc_tz.astimezone(user_tz)
+
+        return final.strftime(fmt)
 
     def _compute_debt_type(self):
         debt_type = self.env["ir.config_parameter"].get_param("pos_debt_notebook.debt_type", default='debt')
@@ -158,7 +209,6 @@ class PosConfig(models.Model):
 
         debt_journal_inactive = journal_obj.search([
             ('code', '=', 'TDEBT'),
-            ('name', '=', 'Debt Journal'),
             ('company_id', '=', user.company_id.id),
             ('debt', '=', False),
         ])
@@ -282,3 +332,32 @@ class PosOrder(models.Model):
             for o_line in order.lines:
                 product_list.append('%s(%s * %s) + ' % (o_line.product_id.name, o_line.qty, o_line.price_unit))
             order.product_list = ''.join(product_list).strip(' + ')
+
+
+class PosCreditUpdate(models.Model):
+    _name = 'pos.credit.update'
+    _description = "Manual Credit Updates"
+
+    partner_id = fields.Many2one('res.partner', string="Partner", required=True)
+    user_id = fields.Many2one(
+        'res.users',
+        string='Salesperson',
+        default=lambda s: s.env.user,
+        readonly=True
+    )
+    company_id = fields.Many2one(
+        'res.company',
+        string='Company',
+        required=True,
+        default=lambda s: s.env.user.company_id,
+    )
+    currency_id = fields.Many2one(
+        'res.currency',
+        string='Currency',
+        default=lambda s: s.env.user.company_id.currency_id,
+    )
+    balance = fields.Monetary('Balance Update', help="Change of balance. Negative value for purchases without money (debt). Positive for credit payments (prepament or payments for debts).")
+    note = fields.Text('Note')
+    date = fields.Datetime(string='Date', default=fields.Date.today, required=True)
+
+    state = fields.Selection([('confirm', 'Confirmed'), ('cancel', 'Canceled')], default='confirm', required=True)