+++ pos_product_category_discount/static/src/js/discount_program.js
@@ -16,31 +16,66 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
 
     var PosModelSuper = models.PosModel;
     models.PosModel = models.PosModel.extend({
-        get_discount_category: function(id) {
-            var self = this;
-            var model = new Model('pos.category_discount');
-            var domain = [['discount_program_id', '=', id]];
-            model.call('search_read', [domain]).then(function (resultat) {
-                resultat.forEach(function(item){
-                    self.apply_discount_category(item.category_discount_pc, item.discount_category_id[0], item.discount_program_id[1]);
-                });
+        initialize: function (session, attributes) {
+            var partner_model = _.find(this.models, function(model){
+                return model.model === 'res.partner';
             });
+            partner_model.fields.push('discount_program_id');
+            return PosModelSuper.prototype.initialize.apply(this, arguments);
         },
-        apply_discount_category: function(discount, category_id, discount_program_name) {
+        get_discount_category: function(id) {
+            if (this.config.iface_discount) {
+                var self = this;
+                var model = new Model('pos.category_discount');
+                var domain = [['discount_program_id', '=', id]];
+                var order = this.get_order();
+                model.call('search_read', [domain]).then(function (resultat) {
+                    order.remove_all_discounts();
+                    resultat.forEach(function(item){
+                        self.apply_discount_category(item);
+                    });
+                });
+            } else {
+                return false;
+            }
+        },
+        apply_discount_category: function(discount_program) {
             var self = this;
             var order = this.get_order();
-            var lines = order.get_orderlines().filter(function (item) { return item.product.pos_categ_id[0] == category_id && item.product.discount_allowed; });
-            lines.forEach(function (item){
-                item.discount = discount;
-                item.discountStr = discount;
-                item.discount_program_name = discount_program_name;
-                order.get_orderline(item.id).set_discount(discount);
+            var lines = order.get_orderlines().filter(function(item) {
+                return item.product.pos_categ_id[0] === discount_program.discount_category_id[0] && item.product.discount_allowed;
             });
+            lines.forEach(function (item){
+                item.discount = discount_program.category_discount_pc;
+                item.discountStr = discount_program.category_discount_pc;
+                item.discount_program_name = discount_program.discount_program_id[1];
+                order.get_orderline(item.id).set_discount(discount_program.category_discount_pc);
+            });
+            order.current_discount_program = discount_program.discount_category_id;
         },
     });
 
     var OrderlineSuper = models.Orderline;
     models.Orderline = models.Orderline.extend({
+        initialize: function(attr,options){
+            OrderlineSuper.prototype.initialize.apply(this,arguments);
+            if (this.order && this.order.current_discount_program) {
+                this.apply_product_discount(this.order.current_discount_program[0]);
+            }
+        },
+        apply_product_discount: function(id) {
+            var self = this;
+            var model = new Model('pos.category_discount');
+            var domain = [['discount_program_id', '=', id]];
+            model.call('search_read', [domain]).then(function (result) {
+                result.forEach(function(res) {
+                    if (res.discount_category_id[0] === self.product.pos_categ_id[0]) {
+                        self.discount_program_name = res.discount_program_id[1];
+                        self.set_discount(res.category_discount_pc);
+                    }
+                });
+            });
+        },
         export_as_JSON: function(){
             var json = OrderlineSuper.prototype.export_as_JSON.call(this);
             json.discount_program_name = this.discount_program_name || false;
@@ -57,14 +92,24 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
 
     var OrderSuper = models.Order;
     models.Order = models.Order.extend({
+        remove_all_discounts: function() {
+            if (this.pos.config.iface_discount) {
+                this.current_discount_program = false;
+                this.get_orderlines().forEach(function(line){
+                    line.set_discount(false);
+                });
+            }
+        },
         export_as_JSON: function(){
             var json = OrderSuper.prototype.export_as_JSON.call(this);
             json.product_discount = this.product_discount || false;
+            json.current_discount_program = this.current_discount_program;
             return json;
         },
         init_from_JSON: function(json) {
             OrderSuper.prototype.init_from_JSON.apply(this,arguments);
             this.product_discount = json.product_discount || false;
+            this.current_discount_program = json.current_discount_program;
         },
     });
 
@@ -83,17 +128,19 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
         update_summary: function(){
             this._super();
             var order = this.pos.get_order();
-            var discount     = order ? order.get_total_discount() : 0;
+            var discount = order
+                           ? order.get_total_discount()
+                           : 0;
             if (this.el.querySelector('.summary .total .discount .value')) {
                 if (order.product_discount) {
-                    discount = discount - order.product_discount;
+                    discount -= order.product_discount;
                 }
                 this.el.querySelector('.summary .total .discount .value').textContent = this.format_currency(discount);
             }
         },
         remove_orderline: function(order_line){
             this._super(order_line);
-            if (order_line.product.id == this.pos.config.discount_product_id[0]) {
+            if (order_line.product.id === this.pos.config.discount_product_id[0]) {
                 var order = this.pos.get_order();
                 order.product_discount = false;
             }
@@ -107,9 +154,9 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
             if (this.gui && this.gui.screen_instances.products && this.gui.screen_instances.products.action_buttons.discount) {
                 var disc_widget = this.gui.screen_instances.products.action_buttons.discount;
                 disc_widget.apply_discount = function(pc) {
-                    var order    = self.pos.get_order();
-                    var lines    = order.get_orderlines();
-                    var product  = self.pos.db.get_product_by_id(self.pos.config.discount_product_id[0]);
+                    var order = self.pos.get_order();
+                    var lines = order.get_orderlines();
+                    var product = self.pos.db.get_product_by_id(self.pos.config.discount_product_id[0]);
                     if (pc === 0) {
                         order.product_discount = 0;
                     }
@@ -120,9 +167,10 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
                             }
                         });
 
-
                         // Product with a prohibited discount
-                        var not_discount_product = order.get_orderlines().filter(function (item) {return item.product.discount_allowed === false;});
+                        var not_discount_product = order.get_orderlines().filter(function(item) {
+                            return item.product.discount_allowed === false;
+                        });
 
                         // Common price without discount for product with a prohibited discount
                         var price_without_discount = 0;
@@ -135,9 +183,8 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
                                 } else {
                                     price = item.price;
                                 }
-                                price_without_discount = price_without_discount + price;
+                                price_without_discount += price;
                             });
-
                         }
                         // Discount
                         var discount = - pc / 100.0 * (order.get_total_with_tax() - price_without_discount);
@@ -155,7 +202,9 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
                         'confirm': function (val) {
                             if (val) {
                                 val = Math.round(Math.max(0, Math.min(100, val)));
-                            } else val = null;
+                            } else {
+                                val = null;
+                            }
                             self.gui.screen_instances.products.action_buttons.discount.apply_discount(val);
                         },
                     });
@@ -193,7 +242,9 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
     models.load_models({
         model: 'pos.discount_program',
         fields: ['discount_program_name', 'discount_program_number', 'id'],
-        domain: function(self){ return [['discount_program_active','=',true]]; },
+        domain: function(self){
+            return [['discount_program_active','=',true]];
+        },
         loaded: function(self,discount_program){
             var sorting_discount_program = function(idOne, idTwo){
                 return idOne.discount_program_number - idTwo.discount_program_number;
@@ -210,7 +261,7 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
         loaded: function(self,products){
             products.forEach(function(item){
                 if (item.product_variant_id) {
-                    var product  = self.db.get_product_by_id(item.product_variant_id[0]);
+                    var product = self.db.get_product_by_id(item.product_variant_id[0]);
                     if (product) {
                         product.discount_allowed = item.discount_allowed;
                     }
@@ -225,8 +276,14 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
             this._super(options);
             this.popup_discount = false;
             if (options.disc_program) {
-                self.popup_discount = true;
-                self.events["click .discount-program-list .button"] = "click_discount_program";
+                this.popup_discount = true;
+                this.events = _.extend(this.events || {}, {
+                    'click .discount-program-list .button': 'click_discount_program',
+                    'click .reset': function() {
+                        self.pos.get_order().remove_all_discounts();
+                        self.gui.close_popup();
+                    },
+                });
             }
         },
         renderElement: function(){
@@ -236,15 +293,14 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
             }
         },
         get_discount_program_by_id: function(id) {
-            var disc_program_by_id = this.options.disc_program.filter(function (item) {
+            return this.options.disc_program.find(function (item) {
                 return item.id === Number(id);
             });
-            return disc_program_by_id[0];
         },
         click_discount_program: function(e) {
             var self = this;
             var id = e.currentTarget.id;
-            if (id == 'other') {
+            if (id === 'other') {
                 self.gui.show_screen('discountlist');
             } else {
                 this.current_disc_program = this.get_discount_program_by_id(id);
@@ -296,7 +352,7 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
             var contents = this.$el[0].querySelector('.discount-list-contents');
             contents.innerHTML = "";
             for(var i = 0, len = Math.min(discounts.length,1000); i < len; i++){
-                var discount    = discounts[i];
+                var discount = discounts[i];
                 var discountline = this.discount_cache.get_node(discount.id);
                 if(!discountline){
                     var discountline_html = QWeb.render('DiscountLine',{widget: this, discount:discounts[i]});
@@ -314,12 +370,12 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
         },
         toggle_save_button: function(){
             var $button = this.$('.button.next');
-            if (!this.show_disc_button) {
-                $button.addClass('oe_hidden');
-                return;
-            } else {
+            if (this.show_disc_button) {
                 $button.removeClass('oe_hidden');
                 $button.text(_t('Apply'));
+            } else {
+                $button.addClass('oe_hidden');
+                return;
             }
         },
         line_select: function(event,$line,id){
@@ -344,4 +400,44 @@ odoo.define('pos_product_category_discount.discount_program', function (require)
         },
     });
     gui.define_screen({name:'discountlist', widget: DiscountProgramScreenWidget});
+
+    gui.Gui.prototype.screen_classes.filter(function(el) {
+        return el.name === 'clientlist';
+    })[0].widget.include({
+        save_changes: function(){
+            var order = this.pos.get_order();
+            if (this.new_client) {
+                if ((this.has_client_changed() || this.has_discount_program_changed()) &&
+                    this.new_client && this.new_client.discount_program_id) {
+                    this.pos.get_discount_category(this.new_client.discount_program_id[0]);
+                }
+            } else {
+                this.pos.get_order().remove_all_discounts();
+            }
+            this._super();
+        },
+        has_discount_program_changed: function(){
+            if (this.old_client && this.new_client && this.old_client.id === this.new_client.id) {
+                if (this.old_client.discount_program_id && this.new_client.discount_program_id && (
+                    this.old_client.discount_program_id[0] === this.new_client.discount_program_id[0]
+                )) {
+                    return false;
+                }
+                return true;
+            }
+        },
+        toggle_save_button: function() {
+            var $button = this.$('.button.next');
+            if (!this.editing_client && this.has_discount_program_changed()) {
+                 $button.text(_t('Apply Change'));
+                 $button.toggleClass('oe_hidden',!this.has_discount_program_changed());
+            } else {
+                this._super();
+            }
+        },
+        saved_client_details: function(partner_id){
+            this.partner_cache.clear_node(partner_id);
+            this._super(partner_id);
+        },
+    });
 });